lab-03-agents/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.5.0
wikipedia>=1.4.0
requests>=2.31.0
duckduckgo-search>=4.1.0

